+++ CHANGELOG.md
@@ -12,9 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 <!-- - Amélioration des performances lors du chargement des données -->
 - Alexnet model weight initialization for best convergence.
+- The dataset transformation for context encoding of JEPA is updated with image region masker.
 
 ### Fixed
 <!-- - Correction d'un bug empêchant l'export dans certains cas -->
+- Fixing of the bugs linked to weight decay definition for Alexnet JEPA training.
 
 ### Deprecated
 <!-- - L'ancienne API v1 est dépréciée et sera supprimée dans la version 2.0 -->